+++ apps/mobile/components/ui/index.ts
@@ -10,5 +10,7 @@ export * from "./switch";
 export * from "./sheet";
 export * from "./dialog";
 export * from "./select";
+export * from "./search-select";
+export * from "./native-select";
 export { toast, Toaster, default as toastDefault } from "./toast";
 export * from "./skeleton";